fix(app): update cowswap to latest, fix ESM build, extract SwapBridgeTabs component
- Bump @cowprotocol/widget-lib 0.22.3→2.0.1, widget-react 1.0.9→1.0.11
- Fix theme API: replace palette object with 'dark' string (v2 requires all colors)
- Add webpack rule to parse @cowprotocol packages as javascript/esm (iframe-transport ships ESM source)
- Extract SwapBridgeTabs + BridgeSection + SwapSection to ~/components/swap/SwapBridgeTabs.tsx
  (Next.js 15 disallows non-standard named exports from page files)
- Fix unnecessary type assertions in swap/page and useBridge (viem/wagmi upgrade)
