+++ packages/app/src/app/bridge/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 
-import { SwapBridgeLayout } from '~/app/swap/page';
+import { SwapBridgeTabs } from '~/components/swap/SwapBridgeTabs';
 
 export default function BridgePage() {
-  return <SwapBridgeLayout activeTab="bridge" />;
+  return <SwapBridgeTabs activeTab="bridge" />;
 }
