+++ packages/app/src/components/swap/CowSwapEmbed.tsx
@@ -28,13 +28,7 @@ export function CowSwapEmbed() {
             tradeType: TradeType.SWAP,
             sell: { asset: '' },
             buy: { asset: 'USDC' },
-            theme: {
-              baseTheme: 'dark',
-              paper: '#0A0A0A',
-              primary: '#65D9E9',
-              background: '#0A0A0A',
-              text: '#FFFFFF',
-            },
+            theme: 'dark',
             hideNetworkSelector: true,
             tokenLists: [`${API_URL}/tokenlist.json`],
           },
